pyproject.toml
[tool.poetry]
name = "-"
version = "1.0.0"
description = "."
authors = ["gadelschina.di@yandex.ru"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.4"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
async-timeout = "5.0.0"
asyncpg = "0.30.0"
certifi = "2024.8.30"
cffi = "1.17.1"
click = "8.1.7"
colorama = "0.4.6"
cryptography = "43.0.3"
deprecated = "1.2.14"
exceptiongroup = "1.2.2"
h11 = "0.14.0"
httpcore = "1.0.6"
httpx = "0.27.2"
idna = "3.10"
importlib-resources = "6.4.5"
iniconfig = "2.0.0"
packaging = "24.1"
pluggy = "1.5.0"
pycparser = "2.22"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pytest = "8.3.3"
pyyaml = "6.0.2"
sniffio = "1.3.1"
starlette = "0.41.2"
tomli = "2.0.2"
typing-extensions = "4.12.2"
uvicorn = "0.32.0"
wrapt = "1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
